# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Modelo de API para transações da Saque e Pague. O parceiro interessado\
    \ deve implementar as operações desejadas para que a SeP consiga se comunicar\
    \ com o mesmo."
  version: "1.19.0"
  title: "Saque e Pague Transacional"
  contact:
    name: "Time Saque e Pague"
    url: "http://www.saqueepague.com.br"
host: "api-saqueepague.sensedia.com"
basePath: "/saqueepague/SepTransaction/1.19.0"
...
schemes:
- "https"
paths:
...
  /compraTitulosCapitalizacao:
    post:
      tags:
      - "Títulos de Capitalização"
      description: "Operação de compra de títulos de capitalização."
      operationId: "compraTitulosCapitalizacaoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de operação de compra de títulos de capitalizaçã\
          o."
        required: true
        schema:
          $ref: "#/definitions/compraTitulosCapitalizacaoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/compraTitulosCapitalizacaoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "TtulosDeCapitalizao"
  /compraTitulosCapitalizacaoConf:
    post:
      tags:
      - "Títulos de Capitalização"
      description: "Confirmação de operação de compra de títulos de capitalização."
      operationId: "compraTitulosCapitalizacaoConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de compra de titulos de\
          \ capitalizacao"
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "TtulosDeCapitalizao"
  /consultaConta:
    post:
      tags:
      - "Depósito sem Cartão"
      - "Depósito via Carteira Digital"
      - "Transferência com Cartão"
      - "Transferência sem Cartão"
      description: "Operação de consulta de conta de favorecido, usada previamente\
        \ ao depósito para verificar se a conta é válida e está habilitada para depó\
        sito."
      operationId: "consultaContaPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta de conta."
        required: true
        schema:
          $ref: "#/definitions/consultaContaReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaContaResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "DepsitoSemCarto"
...
  /consultaTitulosCapitalizacao:
    post:
      tags:
      - "Títulos de Capitalização"
      description: "Operação de consulta dos possíveis produtos de títulos de capitalizaç\
        ão de uma operadora."
      operationId: "consultaTitulosCapitalizacaoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta produtos de títulos de capitalização."
        required: true
        schema:
          $ref: "#/definitions/consultaTitulosCapitalizacaoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaTitulosCapitalizacaoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "TtulosDeCapitalizao"
...
  /transferencia:
    post:
      tags:
      - "Transferência com Cartão"
      - "Transferência sem Cartão"
      description: "Operação de transferência entre contas."
      operationId: "transferenciaPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de operação de transferência."
        required: true
        schema:
          $ref: "#/definitions/transferenciaReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/transferenciaResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "TransfernciaComCarto"
  /transferenciaConf:
    post:
      tags:
      - "Transferência com Cartão"
      - "Transferência sem Cartão"
      description: "Confirmação de operação de transferência."
      operationId: "transferenciaConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de transferência."
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "TransfernciaComCarto"
definitions:
...
  cartao:
    type: "object"
    properties:
      modoEntrada:
        type: "string"
        description: "Modo de entrada de leitura do cartão. (000 = sem cartão, 050\
          \ = chip, 800 = chip, se falhar realiza leitura da trilha, 900 = trilha\
          \ completa)."
        enum:
        - "000"
        - "050"
        - "800"
        - "900"
      pan:
        type: "string"
        example: "zhXTyHl5nf2ckysYXXjAgZS5fzS7ybC5xePnWeLEP//bw=="
        description: "PAN/Número do cartão utilizado (dado criptografado)."
        pattern: "^.{10,}$"
      trilha2:
        type: "string"
        example: "mMIrkIrf2pSHccWuaROxoIS+NGcTOWcTWXUtjniTa1BRzPziiE8pWnknAxW9HIgoKdc/Fli2KZwNA=="
        description: "Trilha 2 do cartão utilizado (dado criptografado)."
        pattern: "^.{10,}$"
  compraTitulosCapitalizacaoReq:
    type: "object"
    required:
    - "InfCompraTitulosCapitalizacao"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfCompraTitulosCapitalizacao:
        $ref: "#/definitions/infCompraTitulosCapitalizacaoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  compraTitulosCapitalizacaoResp:
    type: "object"
    required:
    - "InfCompraTitulosCapitalizacao"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfCompraTitulosCapitalizacao:
        $ref: "#/definitions/infCompraTitulosCapitalizacaoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfCompraTitulosCapitalizacao:
        recibo: "@               TITULOS COMPRADOS                @              \
          \                                  @ NOME DO TITULO:  XXXXXXXXXXXXX    \
          \             @ QUANTIDADE: 2                                  @ VALOR:\
          \ R$ 25,00                                @ VALOR TOTAL: R$ 50,00      \
          \                    @ TELEFONE: 05199999999                          @\
          \ CPF: 02358422785                               "
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
  consultaContaReq:
    type: "object"
    required:
    - "InfConsultaConta"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfConsultaConta:
        $ref: "#/definitions/infConsultaContaReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaContaResp:
    type: "object"
    required:
    - "InfConsultaConta"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfConsultaConta:
        $ref: "#/definitions/infConsultaContaResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      InfConsultaConta:
        solicDoc: "00"
        modalidadeDeposito: "00"
        nomeCliente: "Carl Edward Sagan"
        depositoIdentificado: "00"
        habilitaDeposito: "01"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  consultaTitulosCapitalizacaoReq:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaTitulosCapitalizacaoResp:
    type: "object"
    required:
    - "InfConsultaTitulosCapitalizacao"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfConsultaTitulosCapitalizacao:
        $ref: "#/definitions/infConsultaTitulosCapitalizacaoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      InfConsultaTitulosCapitalizacao:
        produtos:
        - limiteMaximoCompra: "4"
          tema:
            estilos:
            - valor: "#00288C"
              nome: "cor-fundo ou cor-botao"
            - valor: "#00288C"
              nome: "cor-fundo ou cor-botao"
          valor: "000000000500"
          nome: "Grêmio Mais"
          id: "7"
          dataFimVigencia: "20181130"
          dataInicioVigencia: "20181122"
          detalhes: "Edição ouro de mais um produto do Grêmio Mais"
          descricao: "Edição 115 Anos"
        - limiteMaximoCompra: "4"
          tema:
            estilos:
            - valor: "#00288C"
              nome: "cor-fundo ou cor-botao"
            - valor: "#00288C"
              nome: "cor-fundo ou cor-botao"
          valor: "000000000500"
          nome: "Grêmio Mais"
          id: "7"
          dataFimVigencia: "20181130"
          dataInicioVigencia: "20181122"
          detalhes: "Edição ouro de mais um produto do Grêmio Mais"
          descricao: "Edição 115 Anos"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  cripto:
    type: "object"
    properties:
      criptogramaChip:
        type: "string"
        description: "Em caso de cartão com chip, contém o criptograma dele."
      hash:
        type: "string"
        description: "Hash de todos os campos que compõem a transação."
      idPositiva:
        type: "string"
        example: "q0WKgCM+wbXjzlY="
        description: "Identificação positiva informada pelo usuário, caso utilizada,\
          \ com no mínimo 8 dígitos corresponde às posições (botões) em cada tela\
          \ apresentada. Caso a senha tenha menos de oito sílabas, são adicionados\
          \ zeros para as telas não existentes (dado criptografado)."
        pattern: "^.{8,}$"
      respostaCliente:
        $ref: "#/definitions/cripto_respostaCliente"
      senha:
        type: "string"
        example: "q0WKgCM+wbXjzlY="
        description: "Senha numérica criptografada."
      tamanhoSenha:
        type: "string"
        example: "15"
        description: "Tamanho da senha numérica."
        pattern: "^\\d+$"
      templateBiometria:
        type: "string"
        description: "Em caso de biometria, contém o template gerado."
      token:
        type: "string"
        example: "02C47DF604EB43B1"
        description: "Em caso de utilização de token para autenticação, contém o token\
          \ informado pelo usuário."
    description: "Informações relacionadas a criptografia e senhas da requisição.\
      \ Varia dependendo da necessidade de cada transação."
  criptoResp:
    type: "object"
    properties:
      hash:
        type: "string"
        description: "Hash de todos os campos que compõem a transação."
    example:
      hash: "hash"
...
  infCompraTitulosCapitalizacaoReq:
    type: "object"
    required:
    - "cpf"
    - "idProduto"
    - "quantidade"
    - "telefone"
    properties:
      cpf:
        type: "string"
        example: "02358422785"
        description: "Número do CPF do cliente favorecido da transação (11 dígitos)."
        pattern: "^\\d{11}$"
      idProduto:
        type: "string"
        example: "7"
        description: "Produto utilizado na operação."
        pattern: "^\\d+$"
      quantidade:
        type: "string"
        example: "2"
        description: "Quantidade de itens do produtos."
        pattern: "^\\d+$"
      telefone:
        type: "string"
        example: "05199999999"
        description: "Telefone do cliente favorecido da transação (11 dígitos = DDD\
          \ com 0 quando número tem 8 dígitos, sem 0 quando número tem 9 dígitos)."
        pattern: "^\\d{11}$"
    description: "Informações da requisição de compra de títulos de capitalização."
  infCompraTitulosCapitalizacaoResp:
    type: "object"
    required:
    - "recibo"
    properties:
      recibo:
        type: "string"
        example: "@               TITULOS COMPRADOS                @             \
          \                                   @ NOME DO TITULO:  XXXXXXXXXXXXX   \
          \              @ QUANTIDADE: 2                                  @ VALOR:\
          \ R$ 25,00                                @ VALOR TOTAL: R$ 50,00      \
          \                    @ TELEFONE: 05199999999                          @\
          \ CPF: 02358422785                               "
        description: "Recibo com informações da compra. Linhas com 48 posições separadas\
          \ por \"@\"."
        pattern: "^[^@]{0,48}(\\@[^@]{0,48})*$"
    description: "Informações da resposta da requisição de compra de títulos de capitalizaç\
      ão."
    example:
      recibo: "@               TITULOS COMPRADOS                @                \
        \                                @ NOME DO TITULO:  XXXXXXXXXXXXX        \
        \         @ QUANTIDADE: 2                                  @ VALOR: R$ 25,00\
        \                                @ VALOR TOTAL: R$ 50,00                 \
        \         @ TELEFONE: 05199999999                          @ CPF: 02358422785\
        \                               "
  infConsultaContaReq:
    type: "object"
    required:
    - "modalidadeDeposito"
    properties:
      codBanco:
        type: "string"
        example: "237"
        description: "Código do banco utilizado na operação (3 dígitos)."
        pattern: "^\\d{3}$"
      cpfFavorecido:
        type: "string"
        example: "02358422785"
        description: "Número do CPF ou CNPJ do cliente favorecido da transação (11\
          \ dígitos ou 14 dígitos respectivamente)."
        pattern: "^\\d{14}|\\d{11}$"
      modalidadeDeposito:
        type: "string"
        example: "00"
        description: "Indica a modalidade de depósito desejada (2 dígitos, 00 = dinheiro,\
          \ 01 = cheque)."
        enum:
        - "00"
        - "01"
      numAgencia:
        type: "string"
        example: "4029"
        description: "Número da agência utilizada na operação."
        pattern: "^\\d+$"
      numConta:
        type: "string"
        example: "0082348296"
        description: "Número da conta utilizada na operação."
        pattern: "^\\d+$"
      telefoneFavorecido:
        type: "string"
        example: "05199999999"
        description: "Telefone do cliente favorecido da transação (11 dígitos = DDD\
          \ com 0 quando número tem 8 dígitos, sem 0 quando número tem 9 dígitos)."
        pattern: "^\\d{11}$"
    description: "Informações da requisição de consulta de conta do favorecido do\
      \ depósito."
...
  infConsultaTitulosCapitalizacaoResp:
    type: "object"
    properties:
      produtos:
        type: "array"
        description: "Produtos de título de capitalização disponíveis."
        items:
          $ref: "#/definitions/capitalizationBondObj"
    description: "Informações da requisição de consulta de produtos de título de capitalizaç\
      ão."
    example:
      produtos:
      - limiteMaximoCompra: "4"
        tema:
          estilos:
          - valor: "#00288C"
            nome: "cor-fundo ou cor-botao"
          - valor: "#00288C"
            nome: "cor-fundo ou cor-botao"
        valor: "000000000500"
        nome: "Grêmio Mais"
        id: "7"
        dataFimVigencia: "20181130"
        dataInicioVigencia: "20181122"
        detalhes: "Edição ouro de mais um produto do Grêmio Mais"
        descricao: "Edição 115 Anos"
      - limiteMaximoCompra: "4"
        tema:
          estilos:
          - valor: "#00288C"
            nome: "cor-fundo ou cor-botao"
          - valor: "#00288C"
            nome: "cor-fundo ou cor-botao"
        valor: "000000000500"
        nome: "Grêmio Mais"
        id: "7"
        dataFimVigencia: "20181130"
        dataInicioVigencia: "20181122"
        detalhes: "Edição ouro de mais um produto do Grêmio Mais"
        descricao: "Edição 115 Anos"
...
  infTransacaoConf:
    type: "object"
    required:
    - "cdProc"
    - "codOperadora"
    - "dataHora"
    - "dataLocal"
    - "horaLocal"
    - "nsu"
    - "nsuResposta"
    - "valor"
    properties:
      cdProc:
        type: "string"
        example: "029100"
        description: "Igual à requisição e à resposta."
        pattern: "^\\d{6}$"
      codMoeda:
        type: "string"
        example: "986"
        description: "Igual à requisição e à resposta."
        pattern: "^\\d{3}$"
      codOperadora:
        type: "string"
        example: "00000000914"
        description: "Igual à requisição e à resposta."
        pattern: "^\\d{11}$"
      dataHora:
        type: "string"
        example: "1122151032"
        description: "Data e hora da transação (MMDDhhmmss)."
        pattern: "^\\d{10}$"
      dataLocal:
        type: "string"
        example: "1122"
        description: "Igual à requisição e à resposta."
        pattern: "^\\d{4}$"
      horaLocal:
        type: "string"
        example: "151032"
        description: "Igual à requisição e à resposta."
        pattern: "^\\d{6}$"
      nsu:
        type: "string"
        example: "000080247206"
        description: "Igual à requisição e à resposta."
        pattern: "^\\d{12}$"
      nsuResposta:
        type: "string"
        example: "820"
        description: "Igual à resposta."
        pattern: "^\\d+$"
      valor:
        type: "string"
        example: "000000005000"
        description: "Valor efetivado na transação. Em alguns casos, pode ser diferente\
          \ da requisição e da resposta (12 dígitos, incluindo centavos)."
        pattern: "^\\d{12}$"
    description: "Informações sobre a requisição de confirmação de transação."
  infTransacaoReq:
    type: "object"
    required:
    - "cdProc"
    - "codOperadora"
    - "dataHora"
    - "dataLocal"
    - "horaLocal"
    - "nsu"
    properties:
      cdProc:
        type: "string"
        example: "029100"
        description: "Código de processamento da transação, gerado pela Saque e Pague.\
          \ Verificar código específico de cada transação com a Saque e Pague (6 dí\
          gitos)."
        pattern: "^\\d{6}$"
      codMoeda:
        type: "string"
        example: "986"
        description: "Código na tabela de moedas da ISO-4217 (986 = real, 3 dígitos)."
        pattern: "^\\d{3}$"
      codOperadora:
        type: "string"
        example: "00000000914"
        description: "Código definido pela Saque e Pague que identifica o autorizador\
          \ (11 dígitos)."
        pattern: "^\\d{11}$"
      dataContabil:
        type: "string"
        example: "1122"
        description: "Data contábil da transação (MMDD)."
        pattern: "^\\d{4}$"
      dataHora:
        type: "string"
        example: "1122151032"
        description: "Data e hora da transação (MMDDhhmmss)."
        pattern: "^\\d{10}$"
      dataLocal:
        type: "string"
        example: "1122"
        description: "Data local (MMDD)."
        pattern: "^\\d{4}$"
      horaLocal:
        type: "string"
        example: "151032"
        description: "Hora local (hhmmss)."
        pattern: "^\\d{6}$"
      nsu:
        type: "string"
        example: "000080247206"
        description: "Identificador da transação gerado pela Saque e Pague (de 9 dí\
          gitos a 12 dígitos)."
        pattern: "^\\d{12}$"
      valor:
        type: "string"
        example: "000000005000"
        description: "Valor da transação (12 dígitos, incluindo centavos)."
        pattern: "^\\d{12}$"
    description: "Informações genéricas sobre a requisição de transação."
  infTransacaoResp:
    type: "object"
    required:
    - "cdProc"
    - "codOperadora"
    - "dataHora"
    - "dataLocal"
    - "horaLocal"
    - "nsu"
    - "nsuResposta"
    properties:
      cdProc:
        type: "string"
        example: "029100"
        description: "Igual à requisição."
        pattern: "^\\d{6}$"
      codMoeda:
        type: "string"
        example: "986"
        description: "Igual à requisição com exceção da transação de consultaSaque."
        pattern: "^\\d{3}$"
      codOperadora:
        type: "string"
        example: "00000000914"
        description: "Igual à requisição."
        pattern: "^\\d{11}$"
      dataHora:
        type: "string"
        example: "1122151032"
        description: "Data e hora da transação (MMDDhhmmss)."
        pattern: "^\\d{10}$"
      dataLocal:
        type: "string"
        example: "1122"
        description: "Igual à requisição."
        pattern: "^\\d{4}$"
      errorMessage:
        type: "string"
        example: ""
        description: "Caso ocorra erro, este campo preenchido identifica que um erro\
          \ ocorreu e seu valor."
      horaLocal:
        type: "string"
        example: "151032"
        description: "Igual à requisição."
        pattern: "^\\d{6}$"
      mensagemCliente:
        type: "string"
        example: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em tela."
        description: "Mensagem para ser mostrada na tela para o cliente, em caso de\
          \ erro (objeto errorMessage deve estar preenchido para identificação de\
          \ um erro)."
      nsu:
        type: "string"
        example: "000080247206"
        description: "Igual à requisição."
        pattern: "^\\d{12}$"
      nsuResposta:
        type: "string"
        example: "820"
        description: "Identificador da resposta da transação, gerado pelo autorizador."
        pattern: "^\\d+$"
      valor:
        type: "string"
        example: "5000"
        description: "Igual à requisição com exceção da transação de consultaSaldo\
          \ e consultaSaque."
        pattern: "^\\d+$"
    description: "Informações sobre a resposta da requisição de transação."
    example:
      cdProc: "029100"
      mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
        \ tela."
      nsu: "000080247206"
      codMoeda: "986"
      codOperadora: "00000000914"
      dataLocal: "1122"
      errorMessage: ""
      valor: "5000"
      horaLocal: "151032"
      nsuResposta: "820"
      dataHora: "1122151032"
...
  transferenciaReq:
    type: "object"
    required:
    - "InfTransacao"
    - "InfTransferencia"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfTransferencia:
        $ref: "#/definitions/infTransferenciaReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  transferenciaResp:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      Inftransferencia:
        $ref: "#/definitions/infTransferenciaResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      Inftransferencia:
        recibo: "          COMPROVANTE DE TRANSFERENCIA          @           TRANSFERENCIA\
          \ ENTRE CONTAS           @ORIGEM: CONTA CORRENTE - 123/1234567890      \
          \   @DESTINO: CONTA CORRENTE - 456/9876543210        @VALOR TRANSFERENCIA:\
          \ 123,456                    "
        origem:
          nome: "Jean-Luc Picard"
        destino:
          nome: "William Thomas Riker"
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
  terminal:
    type: "object"
    required:
    - "codEstab"
    - "id"
    - "tipo"
    properties:
      codEstab:
        type: "string"
        example: "000000000742673"
        description: "Código do estabelecimento onde está localizado o terminal gerado\
          \ pela Saque e Pague (15 dígitos)."
        pattern: "^\\d{15}$"
      id:
        type: "string"
        example: "05100004"
        description: "Identificador do terminal gerado pela Saque e Pague (8 dígitos)."
        pattern: "^\\d{8}$"
      tipo:
        type: "string"
        example: "008"
        description: "Tipo de terminal o qual a operação é gerada (3 dígitos, 008\
          \ = ATM)."
        pattern: "^\\d{3}$"
    description: "Informações sobre o terminal vigente."
    example:
      codEstab: "000000000742673"
      tipo: "008"
      id: "05100004"
  transacConf:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoConf"
      Terminal:
        $ref: "#/definitions/terminal"
...